Refactor gameplay folders
